Modified the indpendent trial proper
Modified the indpendent trial proper
And saving across all the data independent of the trials. Also used this
file to record from the subject yesterday successfully. Made some small
mistakes. (1) Although I added new lines in the count 1 case but not in
the more than count 1 case.  (2) The inrefact as now is inside the
trials, it has become trial specific now, So I had to include the same
in the case of  constant case trials also.
Future Improvement: In future this can be improved by keeping this same
repited codes inside some function.